refactor: move the fake LLM clients into tests/helpers
FakeLLMClient and RecordingClient are about to be shared with the
tests for the grammar-guided LLM transformation pumps.
